Split up browse/test-utils
