+++ frontend/src/metabase/browse/models/test-utils.ts
@@ -4,12 +4,7 @@ import {
   createMockSearchResult,
 } from "metabase-types/api/mocks";
 
-import type {
-  MetricResult,
-  ModelResult,
-  RecentMetric,
-  RecentModel,
-} from "./types";
+import type { ModelResult, RecentModel } from "./types";
 
 export const createMockModelResult = (
   model: Partial<ModelResult> = {},
@@ -20,16 +15,3 @@ export const createMockRecentModel = (
   model: Partial<RecentCollectionItem>,
 ): RecentModel =>
   createMockRecentCollectionItem({ ...model, model: "dataset" }) as RecentModel;
-
-export const createMockMetricResult = (
-  metric: Partial<MetricResult> = {},
-): MetricResult =>
-  createMockSearchResult({ ...metric, model: "metric" }) as MetricResult;
-
-export const createMockRecentMetric = (
-  metric: Partial<RecentMetric>,
-): RecentMetric =>
-  createMockRecentCollectionItem({
-    ...metric,
-    model: "metric",
-  }) as RecentMetric;
